Backend/requirements.txt
Flask
psycopg2-binary
google-cloud-storage
Flask-Cors
python-dotenv
